+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to Toisto will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.38.1 - [Unreleased]
+
+### Fixed
+
+- Don't show an answer as another correct answer if the user's answer was a generated alternative. Fixes [#1136](https://github.com/fniessink/toisto/issues/1136).
+
 ## 0.38.0 - 2025-09-14
 
 ### Fixed